Add files via upload
This project will focus on using Natural Language Processing (NLP) techniques to find broad trends in the written thoughts of the customers. The goal in this project is to predict whether customers recommend the product they purchased using the information in their review text.
One of the challenges in this project is to extract useful information from the "Review Text" variable using text mining techniques. The other challenge is that we need to convert text files into numeric feature vectors to run machine learning algorithms.